# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TrackVault API",
        "description": "Data Collection and Payment Management System - Complete REST API Documentation",
        "contact": {
            "name": "TrackVault Support",
            "email": "admin@trackvault.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Development Server"
        },
        {
            "url": "https://api.trackvault.com/api",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/api/payments": {
            "get": {
                "tags": [
                    "Payments"
                ],
                "summary": "List all payments",
                "description": "Get paginated list of payments with filtering and sorting",
                "operationId": "a67c4dcf277fa699515fd3b6d4835388",
                "parameters": [
                    {
                        "name": "supplier_id",
                        "in": "query",
                        "description": "Filter by supplier ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "payment_type",
                        "in": "query",
                        "description": "Filter by payment type",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "advance",
                                "partial",
                                "full"
                            ]
                        }
                    },
                    {
                        "name": "from_date",
                        "in": "query",
                        "description": "Filter from date (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to_date",
                        "in": "query",
                        "description": "Filter to date (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Sort by field",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "payment_date",
                            "enum": [
                                "payment_date",
                                "amount",
                                "payment_type",
                                "created_at",
                                "updated_at"
                            ]
                        }
                    },
                    {
                        "name": "sort_order",
                        "in": "query",
                        "description": "Sort order",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "desc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Results per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15,
                            "maximum": 100
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Create a new payment",
                "description": "Create a new payment record",
                "operationId": "6b47fbea138d1f528831ed3e92ee46fd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "supplier_id",
                                    "payment_date",
                                    "amount",
                                    "payment_type"
                                ],
                                "properties": {
                                    "supplier_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "payment_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-12-25"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "minimum": 0.01,
                                        "example": 5000
                                    },
                                    "payment_type": {
                                        "type": "string",
                                        "enum": [
                                            "advance",
                                            "partial",
                                            "full"
                                        ],
                                        "example": "partial"
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "maxLength": 100,
                                        "example": "Bank Transfer"
                                    },
                                    "reference_number": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "PAY-001"
                                    },
                                    "notes": {
                                        "type": "string"
                                    },
                                    "metadata": {
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Payment created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your bearer token in the format: Bearer {token}",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}